add static_casts to resolve warnings

+++ src/mongocxx/gridfs/downloader.cpp
@@ -168,7 +168,7 @@ void downloader::fetch_chunk() {
     _get_impl().chunk_buffer_len = binary_data.size;
 
     if (not _get_impl().chunks_seen) {
-        _get_impl().chunk_buffer_offset = _get_impl().start.bytes_offset;
+        _get_impl().chunk_buffer_offset = static_cast<std::size_t>(_get_impl().start.bytes_offset);
         _get_impl().chunks_seen = chunks_seen;
     } else {
         _get_impl().chunk_buffer_offset = 0;
